Move QlobberSub into qlobber-sub.js
